refactor: organize help request module layers
