+++ src/main/java/com/helpunker/common/exception/BusinessRuleException.java
@@ -1,0 +1,8 @@
+package com.helpunker.common.exception;
+
+public class BusinessRuleException extends RuntimeException {
+
+    public BusinessRuleException(String message) {
+        super(message);
+    }
+}
