+++ src/main/java/com/helpunker/common/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.helpunker.common.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
