--- src/main/java/com/helpunker/domain/model/OutboxEvent.java
@@ -1,46 +1,0 @@
-package com.helpunker.domain.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.OffsetDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "outbox_events")
-public class OutboxEvent {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "aggregate_type", nullable = false, length = 64)
-    private String aggregateType;
-
-    @Column(name = "aggregate_id", nullable = false, length = 64)
-    private String aggregateId;
-
-    @Column(name = "event_type", nullable = false, length = 64)
-    private String eventType;
-
-    @Column(nullable = false, columnDefinition = "jsonb")
-    private String payload;
-
-    @Column(name = "created_at", nullable = false, insertable = false, updatable = false)
-    private OffsetDateTime createdAt;
-
-    @Column(name = "processed_at")
-    private OffsetDateTime processedAt;
-}
