--- src/main/java/com/helpunker/domain/model/RequestStatus.java
@@ -1,9 +1,0 @@
-package com.helpunker.domain.model;
-
-public enum RequestStatus {
-    OPEN,
-    ASSIGNED,
-    ONGOING,
-    COMPLETED,
-    CANCELLED
-}
