--- src/main/java/com/helpunker/domain/model/UserRole.java
@@ -1,7 +1,0 @@
-package com.helpunker.domain.model;
-
-public enum UserRole {
-    ELDERLY,
-    VOLUNTEER,
-    ADMIN
-}
