--- src/main/java/com/helpunker/domain/repository/OutboxEventRepository.java
@@ -1,7 +1,0 @@
-package com.helpunker.domain.repository;
-
-import com.helpunker.domain.model.OutboxEvent;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OutboxEventRepository extends JpaRepository<OutboxEvent, Long> {
-}
