--- src/main/java/com/helpunker/domain/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.helpunker.domain.repository;
-
-import com.helpunker.domain.model.User;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-
-    Optional<User> findByEmail(String email);
-
-    Optional<User> findByPhone(String phone);
-}
