+++ src/main/java/com/helpunker/modules/helprequest/domain/model/HelpRequest.java
@@ -1,0 +1,81 @@
+package com.helpunker.modules.helprequest.domain.model;
+
+import com.helpunker.modules.user.domain.model.User;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import jakarta.persistence.Version;
+import java.math.BigDecimal;
+import java.time.OffsetDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "requests")
+public class HelpRequest {
+
+    @Id
+    private UUID id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "elderly_id", nullable = false)
+    private User elderly;
+
+    @Column(nullable = false, length = 160)
+    private String title;
+
+    @Column(nullable = false)
+    private String details;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private RequestStatus status;
+
+    private String category;
+
+    @Column(name = "location_lat", precision = 9, scale = 6)
+    private BigDecimal locationLat;
+
+    @Column(name = "location_lng", precision = 9, scale = 6)
+    private BigDecimal locationLng;
+
+    private String address;
+
+    @Column(name = "created_at", nullable = false, insertable = false, updatable = false)
+    private OffsetDateTime createdAt;
+
+    @Column(name = "updated_at", nullable = false, insertable = false, updatable = false)
+    private OffsetDateTime updatedAt;
+
+    @Version
+    private long version;
+
+    @Builder.Default
+    @OneToMany(mappedBy = "request", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<RequestPhoto> photos = new ArrayList<>();
+
+    public void addPhoto(RequestPhoto photo) {
+        photos.add(photo);
+        photo.setRequest(this);
+    }
+}
