+++ src/main/java/com/helpunker/modules/helprequest/domain/model/RequestStatus.java
@@ -1,0 +1,9 @@
+package com.helpunker.modules.helprequest.domain.model;
+
+public enum RequestStatus {
+    OPEN,
+    ASSIGNED,
+    ONGOING,
+    COMPLETED,
+    CANCELLED
+}
