+++ src/main/java/com/helpunker/modules/helprequest/infrastructure/sse/RequestEvent.java
@@ -1,0 +1,6 @@
+package com.helpunker.modules.helprequest.infrastructure.sse;
+
+import com.helpunker.modules.helprequest.application.dto.response.HelpRequestResponse;
+
+public record RequestEvent(RequestEventType type, HelpRequestResponse payload) {
+}
