+++ src/main/java/com/helpunker/modules/user/domain/model/User.java
@@ -1,0 +1,54 @@
+package com.helpunker.modules.user.domain.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.persistence.Version;
+import java.time.OffsetDateTime;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "users")
+public class User {
+
+    @Id
+    private UUID id;
+
+    private String phone;
+
+    private String email;
+
+    @Column(name = "display_name", nullable = false)
+    private String displayName;
+
+    @Enumerated(EnumType.STRING)
+    private UserRole role;
+
+    @Column(name = "volunteer_verified")
+    private boolean volunteerVerified;
+
+    @Column(name = "password_hash")
+    private String passwordHash;
+
+    @Column(name = "created_at", nullable = false, insertable = false, updatable = false)
+    private OffsetDateTime createdAt;
+
+    @Column(name = "updated_at", nullable = false, insertable = false, updatable = false)
+    private OffsetDateTime updatedAt;
+
+    @Version
+    private long version;
+}
