+++ src/main/java/com/helpunker/modules/user/domain/model/UserRole.java
@@ -1,0 +1,7 @@
+package com.helpunker.modules.user.domain.model;
+
+public enum UserRole {
+    ELDERLY,
+    VOLUNTEER,
+    ADMIN
+}
