+++ src/main/java/com/helpunker/modules/user/domain/repository/DeviceTokenRepository.java
@@ -1,0 +1,11 @@
+package com.helpunker.modules.user.domain.repository;
+
+import com.helpunker.modules.user.domain.model.DeviceToken;
+import java.util.List;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DeviceTokenRepository extends JpaRepository<DeviceToken, UUID> {
+
+    List<DeviceToken> findByUserId(UUID userId);
+}
