+++ src/main/java/com/helpunker/modules/user/domain/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.helpunker.modules.user.domain.repository;
+
+import com.helpunker.modules.user.domain.model.User;
+import java.util.Optional;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+
+    Optional<User> findByEmail(String email);
+
+    Optional<User> findByPhone(String phone);
+}
