--- src/main/java/com/helpunker/service/CreateHelpRequestCommand.java
@@ -1,19 +1,0 @@
-package com.helpunker.service;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.UUID;
-
-public record CreateHelpRequestCommand(
-        UUID elderlyId,
-        String title,
-        String details,
-        String category,
-        BigDecimal locationLat,
-        BigDecimal locationLng,
-        String address,
-        List<Photo> photos) {
-
-    public record Photo(String url, String contentType) {
-    }
-}
