--- src/main/java/com/helpunker/service/HelpRequestSearchCriteria.java
@@ -1,8 +1,0 @@
-package com.helpunker.service;
-
-import com.helpunker.domain.model.RequestStatus;
-import java.util.UUID;
-
-public record HelpRequestSearchCriteria(
-        RequestStatus status, UUID elderlyId, Double latitude, Double longitude, Double radiusKm) {
-}
