--- src/main/java/com/helpunker/service/exception/BusinessRuleException.java
@@ -1,8 +1,0 @@
-package com.helpunker.service.exception;
-
-public class BusinessRuleException extends RuntimeException {
-
-    public BusinessRuleException(String message) {
-        super(message);
-    }
-}
