--- src/main/java/com/helpunker/service/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.helpunker.service.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
