--- src/main/java/com/helpunker/service/sse/RequestEvent.java
@@ -1,6 +1,0 @@
-package com.helpunker.service.sse;
-
-import com.helpunker.web.response.HelpRequestResponse;
-
-public record RequestEvent(RequestEventType type, HelpRequestResponse payload) {
-}
