--- src/main/java/com/helpunker/service/sse/RequestEventType.java
@@ -1,7 +1,0 @@
-package com.helpunker.service.sse;
-
-public enum RequestEventType {
-    REQUEST_CREATED,
-    REQUEST_UPDATED,
-    REQUEST_CANCELLED
-}
