--- src/main/java/com/helpunker/web/response/PagedResponse.java
@@ -1,6 +1,0 @@
-package com.helpunker.web.response;
-
-import java.util.List;
-
-public record PagedResponse<T>(List<T> content, int page, int size, long totalElements, int totalPages) {
-}
