--- src/main/java/com/helpunker/web/response/RequestPhotoResponse.java
@@ -1,6 +1,0 @@
-package com.helpunker.web.response;
-
-import java.util.UUID;
-
-public record RequestPhotoResponse(UUID id, String url, String contentType) {
-}
